+++ src/lexer/lexer.ts
@@ -106,9 +106,11 @@ export class Lexer {
 	}
 
 	private token(context: Context, type: TokenType, data?: Array<string>): Token {
+		const size = data ? data[0].length : undefined;
+		const location = context.getLocation(size);
 		return {
 			type,
-			location: context.getLocation(),
+			location,
 			data: data ? [].concat(data) : null,
 		};
 	}